Actualiza la template

+++ template.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-#define forr(i, a, b) for(int i = (a); i < (b); i++)
+#define forr(i, a, b) for(int i = (a); i < (int) (b); i++)
 #define forn(i, n) forr(i, 0, n)
 #define forall(it, v) for(auto it = v.begin(); it != v.end(); ++it)
 #define dforn(i, n) for(int i = n - 1; i >= 0; i--)
